+++ Vault/Changes.md
@@ -6,8 +6,15 @@
   - Added `.length` for string length-based conditions
 - **ConnectionPool**
   - Added `.logging(...)`
+- **FromRowFactory**
+  - Added `.getIncluding(...)` and `.getWithColumn(...)`
+- **ManyRowModelAccess**
+  - Added `.pullWithColumn(...)` and `.groupByColumn(...)`
 - **References**
   - Added `.replace(...)` for replacing all references made to a database row
+### Other changes
+- **FromResultFactory**'s `.expandedSelect(...)` now supports inclusion of additional selected items 
+  (columns, tables, etc.)
 
 ## v1.20 - 04.10.2024
 This update adds support for custom selection targets